Version Agnostic Updating The scripts will now download the latest full release from GitHub, Regardless of Version.
This negates the need for manual updating of version numbers within all scripts.

Also switch to Curl for downloading the file since it looks nicer.

Please note; due to the fact that this uses the GitHub API. We have a limited amount of requests that we can make.
I will be implementing some logic to check if this is the case and notify the user.
Alternatively; we can poll GitHub for the rate to check we are able to submit a core request before polling for the version. This will avoid blocking on the server/client side