Implement Updating Implement automatic updating (using the latest release on GitHub. This is fetched using their API and is limited to 60 requests an hour.
The script will also check the currently running Paycoin binary; and if it is the same as the current release it will not update. This can be bypassed using the -f flag, which will cause Paycoin to be stopped and updated regardless of which version it is running.

The updater script supports the following flags
-h                     Show this message
-f                     Stop and force update Paycoin
-r                     Reboot without prompting for interaction
-v                     Show latest available version
You can use the -r and -f flags in conjunction with crontab to schedule automatic reboots and updates.

All scripts by default will now prompt for user interaction before rebooting the machine.
This will allow for more custom setups